small authentication fix for INSTALL.php

+++ include/access_rules.php
@@ -86,11 +86,11 @@
     # automated delete items permission
     $query = 'SELECT config_class FROM ConfigClasses WHERE nav_privs = "user"';
     $user_class_permissions = db_handler($query, "array_direct", "Select all classes where user has permission");
-    
-    foreach ($user_class_permissions AS $permit_class){
-        $NConf_PERMISSIONS->setURL('delete_item.php',       FALSE, array('user'),   array('item' => $permit_class) );
-    }    
-
+    if ($user_class_permissions){
+        foreach ($user_class_permissions AS $permit_class){
+            $NConf_PERMISSIONS->setURL('delete_item.php',       FALSE, array('user'),   array('item' => $permit_class) );
+        }
+    }
 
     # Hosts Service view
     $NConf_PERMISSIONS->setURL('modify_item_service.php', FALSE, array('user') );
@@ -112,4 +112,7 @@
 
     # Update process for admin
     $NConf_PERMISSIONS->setURL('UPDATE.php', TRUE, array('admin') );
+    
+    # Install for all
+    $NConf_PERMISSIONS->setURL('INSTALL.php');
 ?>
